Check of the file already exists

+++ scripts/cadc_query.py
@@ -134,10 +134,13 @@ if __name__ == '__main__':
             try:
                 import wget
                 wget = wget.download
+                print("Using the Python wget package")
             except:
                 print("WARNING: Install the 'wget' package (will use the wget system for now)")
                 wget = lambda url: os.system("wget -nc %s" % url)
             for url in urls:
+                if os.path.exist(url.split('/')[-1] + '.fits.fz')):
+                    continue
                 print("\nDownloading", url)
                 wget(url)
                 os.rename(url.split('/')[-1], url.split('/')[-1] + '.fits.fz')
